# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Architect
  version: 0.1.0
paths:
...
  /memory/short:
    get:
      summary: Get Short Memory
      description: 'Retrieve conversation turns and summary for user/session.


        Requires analyst role. Returns empty on disabled flag. Includes pruned count.'
      operationId: get_short_memory_memory_short_get
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryShortResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Short Memory
      description: 'Clear turns and summary for user/session.


        Requires analyst role. Idempotent: returns cleared=true when enabled.'
      operationId: delete_short_memory_memory_short_delete
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Short Memory Memory Short Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/long:
    get:
      summary: Get Long Memory
      description: 'Retrieve facts for user, optionally filtered by query.


        Requires analyst role. Returns empty on disabled flag.'
      operationId: get_long_memory_memory_long_get
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryLongResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Long Memory
      description: 'Clear all facts for user.


        Requires analyst role. Idempotent: returns cleared=true when enabled.'
      operationId: delete_long_memory_memory_long_delete
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Long Memory Memory Long Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /memory/long/export:
    get:
      summary: Export Long Memory
      description: 'Export all facts for user with retention pruning applied.


        Requires analyst role. Includes embedding presence metadata.'
      operationId: export_long_memory_memory_long_export_get
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryLongResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /memory/long/import:
    post:
      summary: Import Long Memory
      description: 'Ingest facts for user with automatic eviction on size limits.


        Requires analyst role. Tracks evictions triggered during import.'
      operationId: import_long_memory_memory_long_import_post
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryImportPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Import Long Memory Memory Long Import Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MemoryImportPayload:
      properties:
        facts:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Facts
      type: object
      required:
      - facts
      title: MemoryImportPayload
    MemoryLongResponse:
      properties:
        facts:
          items: {}
          type: array
          title: Facts
        audit:
          additionalProperties: true
          type: object
          title: Audit
      type: object
      required:
      - facts
      - audit
      title: MemoryLongResponse
    MemoryShortResponse:
      properties:
        turns:
          items: {}
          type: array
          title: Turns
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        audit:
          additionalProperties: true
          type: object
          title: Audit
      type: object
      required:
      - turns
      - audit
      title: MemoryShortResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError